pyproject.toml
[tool.poetry]
name = "job-fit-enhancer"
version = "0.1.0"
description = ""
authors = ["RegiReis7 <reginaldo@live.com.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
crewai = {extras = ["tools"], version = "^0.30.11"}
streamlit = "^1.35.0"
pypdf2 = "^3.0.1"
langchain-mistralai = "^0.1.7"
boto3 = "^1.34.113"
duckduckgo-search = "^6.1.1"
faiss-cpu = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
